fix: db reconcilation
